src/pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic (>=1.14.1,<2.0.0)",
    "aniso8601 (>=10.0.0,<11.0.0)",
    "attrs (>=25.1.0,<26.0.0)",
    "blinker (>=1.9.0,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "importlib-resources (>=6.5.2,<7.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "jsonschema-specifications (>=2024.10.1,<2025.0.0)",
    "mako (>=1.3.8,<2.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "marshmallow (>=3.26.0,<4.0.0)",
    "marshmallow-sqlalchemy (>=1.4.0,<2.0.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "packaging (>=24.2,<25.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytz (>=2024.2,<2025.0)",
    "referencing (>=0.36.2,<0.37.0)",
    "rpds-py (>=0.22.3,<0.23.0)",
    "six (>=1.17.0,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "tzdata (>=2025.1,<2026.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
